Refactor common code between ssh and ca to cautils and flags.
