+++ cmd/step/main.go
@@ -26,6 +26,7 @@ import (
 	_ "github.com/smallstep/cli/command/fileserver"
 	_ "github.com/smallstep/cli/command/oauth"
 	_ "github.com/smallstep/cli/command/path"
+	_ "github.com/smallstep/cli/command/ssh"
 
 	// Profiling and debugging
 	_ "net/http/pprof"
